requirements.txt
fastapi
uvicorn[standard]
databases
sqlalchemy
pydantic-settings
alembic
fastapi-pagination
psycopg2-binary
